requirements.txt
websocket-client
Flask
Gunicorn
rel
newrelic
